Use OtsuThesholder for deskewing. Fix compilation problems. Add page splitter for double page scans.

+++ pnm_java/src/image/pnm/Pgm.java
@@ -152,7 +152,7 @@ public class Pgm
         return this.cols;
     }
 
-    public int getMaxval()
+    public short getMaxval()
     {
         return this.maxval;
     }
